requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
sqlalchemy==2.0.27
pydantic==2.6.1
python-dotenv==1.0.1
openai==1.12.0
supabase==2.3.5
python-jose==3.3.0
python-multipart==0.0.9
typing_extensions==4.9.0 